+++ Modules/Library/Resources/lang/en/library.php
@@ -30,6 +30,8 @@ return [
     'book_is_overdue' => 'Book is overdue.',
     'book_is_overdue_soon' => 'Book is overdue soon.',
     'book_is_lent_to_person_until' => 'Book is lent to <a href=":route">:person</a> until <strong>:until</strong>.',
+    'book_is_lent_to_soft_deleted_person_until' => 'Book is lent to <strong>:person</strong> (<em>deleted</em>) until <strong>:until</strong>.',
+    'book_is_lent_to_deleted_person_until' => 'Book is lent until <strong>:until</strong> but the person has been deleted from the database.',
     'book_extended' => 'Book extended.',
     'num_books_lending_to_num_persons' => '<strong>:persons</strong> persons are lending <strong>:books</strong> books.',
     'register_new_book' => 'Register new book',